Fix 32-bit hints

+++ forms/mainform.pas
@@ -4314,7 +4314,7 @@ begin
     NewHeight := formPopupTrayslate.MemoTarget.GetTextHeight;
 
     // Add top/bottom padding + controls
-    NewHeight := NewHeight + formPopupTrayslate.PanelPairs.Height + Round(10 * formPopupTrayslate.MemoTarget.ZoomFactor);
+    NewHeight := NewHeight + formPopupTrayslate.PanelPairs.Height + integer(Round(10 * formPopupTrayslate.MemoTarget.ZoomFactor));
 
     // Limit height
     if NewHeight > MaxH then
